OpenIGTLink: create a qmake project
A build system where the supposed way to consume a library from an
application is so broken on Windows that it requires manual directory
specifications to get working… qmake might just be worse than CMake.

And yes, this codebase throws enough integer truncation warnings that
we just disable them all.