Qt: UI: persist probe certificates in the settings file
These actually stay rather consistent across multiple consecutive runs.
It's definitely not necessary to phone home to Clarius every time we
try to run the app, and we can definitely use the app in offline
scenarios if we download the certificates before.
Using a separate per-model `crt` key here to leave room for later
adding the model.